重构了代码
