+++ Makefile
@@ -10,7 +10,7 @@ CFLAGS		= -l $(LIB) -L $(LIBDIR) $(CBIT_FLAG) -Wl,-rpath,'$$ORIGIN/$(LIBDIR_)' -
 CDLL_FLAGS	=  $(CBIT_FLAG) -fPIC -shared -O3 
 
 $(BUILDDIR)/crypto : $(LIBDIR)/libSHA-1_x64.so src/crypto.c
-	$(CC) src/crypto.c $(CFLAGS) -o $(BUILDDIR)/crypto
+	$(CC) src/str.c src/CLI.c src/crypto.c $(CFLAGS) -o $(BUILDDIR)/crypto
 
 $(LIBDIR)/libSHA-1_x64.so : src/SHA-1_x64.c
 	$(CC) src/SHA-1_x64.c $(CDLL_FLAGS) -o $(LIBDIR)/libSHA-1_x64.so
