+++ include/CLI.h
@@ -1,0 +1,1 @@
+int SHA1_CLI(unsigned int argc, char *argv[]);
