+++ include/ERROR.h
@@ -1,0 +1,3 @@
+#define ERROR1 "请输入算法和参数\n"
+#define ERROR2 "请输入参数\n"
+#define ERROR3 "不支持 %s 算法\n"
