+++ include/str.h
@@ -1,0 +1,2 @@
+unsigned long str_len(char str[]);
+char str_cmp(char str0[],char str1[]);
